# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBodyOut": {
        "description": "The body of an error: a top-level code, its params, and optional per-field detail.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/FieldErrorOut"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorBodyOut",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "`{ \"error\": { \"code\": ..., \"params\": {...}, \"fields\": {...}? } }`.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyOut"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FieldErrorOut": {
        "description": "One field's `{code, params}` failure.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "FieldErrorOut",
        "type": "object"
      },
      "ItemEnvelope": {
        "description": "`{ \"item\": Item }`.",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/ItemOut"
          }
        },
        "required": [
          "item"
        ],
        "title": "ItemEnvelope",
        "type": "object"
      },
      "ItemListEnvelope": {
        "description": "`{ \"items\": [Item] }`.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ItemListEnvelope",
        "type": "object"
      },
      "ItemOut": {
        "description": "Wire representation of a line item.",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Amount"
          },
          "country_id": {
            "title": "Country Id",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "currency_code": {
            "title": "Currency Code",
            "type": "string"
          },
          "currency_id": {
            "title": "Currency Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "lat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "map_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "occurred_at": {
            "title": "Occurred At",
            "type": "string"
          },
          "payer_id": {
            "title": "Payer Id",
            "type": "integer"
          },
          "split": {
            "$ref": "#/components/schemas/SplitOut"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "note",
          "occurred_at",
          "currency_code",
          "currency_id",
          "amount",
          "payer_id",
          "country_id",
          "labels",
          "map_url",
          "lat",
          "lon",
          "split",
          "created_at",
          "updated_at"
        ],
        "title": "ItemOut",
        "type": "object"
      },
      "ItemWrite": {
        "additionalProperties": false,
        "description": "Request body for creating or updating a line item.",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "country_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Id"
          },
          "currency_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency Id"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "lat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "map_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map Url"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "payer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payer Id"
          },
          "shares": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares"
          },
          "split_mode": {
            "anyOf": [
              {
                "enum": [
                  "equal",
                  "shares",
                  "exact"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Mode"
          }
        },
        "title": "ItemWrite",
        "type": "object"
      },
...
      "PreviewSplitOut": {
        "description": "Wire representation of a split computed from a request body, without saving.",
        "properties": {
          "split": {
            "$ref": "#/components/schemas/SplitOut"
          },
          "total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "split",
          "total"
        ],
        "title": "PreviewSplitOut",
        "type": "object"
      },
      "PreviewSplitRequest": {
        "description": "Takes the same body an item write does; only these fields are read.\n\n(API.md \u00a73) \u2014 extra fields (name, payer_id, ...) are ignored, not rejected.",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "currency_id": {
            "title": "Currency Id",
            "type": "integer"
          },
          "shares": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares"
          },
          "split_mode": {
            "default": "equal",
            "enum": [
              "equal",
              "shares",
              "exact"
            ],
            "title": "Split Mode",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency_id"
        ],
        "title": "PreviewSplitRequest",
        "type": "object"
      },
...
      "SplitOut": {
        "description": "Wire representation of an item's split across its shares.",
        "properties": {
          "mode": {
            "enum": [
              "equal",
              "shares",
              "exact"
            ],
            "title": "Mode",
            "type": "string"
          },
          "shares": {
            "items": {
              "$ref": "#/components/schemas/ShareOut"
            },
            "title": "Shares",
            "type": "array"
          }
        },
        "required": [
          "mode",
          "shares"
        ],
        "title": "SplitOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Expense tracking for a shared trip: people, currencies, countries and labels belong to a trip; every line item records who paid and how the cost splits across the roster.\n\nJSON in, JSON out. Every response is a single object with its payload under a named key, never a bare array. Amounts are plain numbers in the currency's own units - typed values to 2 fraction digits, computed shares to 6 - and are taken on input as canonical decimal strings (`\"18400.50\"`). Nothing is ever converted between currencies: balances and statistics are per currency.\n\nErrors are `{code, params}` with no message text, so a client renders them through its own catalog in its own language.",
    "title": "Trip Accounter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trips/{slug}/items": {
      "get": {
        "description": "List items for a trip.",
        "operationId": "list_items_api_v1_trips__slug__items_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Items",
        "tags": [
          "items"
        ]
      },
      "post": {
        "description": "Create a new item on a trip.",
        "operationId": "create_item_api_v1_trips__slug__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Item",
        "tags": [
          "items"
        ]
      }
    },
    "/api/v1/trips/{slug}/items/preview-split": {
      "post": {
        "description": "Preview how an amount would split without creating an item.",
        "operationId": "preview_split_api_v1_trips__slug__items_preview_split_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewSplitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewSplitOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Preview Split",
        "tags": [
          "items"
        ]
      }
    },
    "/api/v1/trips/{slug}/items/{item_id}": {
      "delete": {
        "description": "Delete an item from a trip.",
        "operationId": "delete_item_api_v1_trips__slug__items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete Item",
        "tags": [
          "items"
        ]
      },
      "get": {
        "description": "Get a single item by id.",
        "operationId": "get_item_api_v1_trips__slug__items__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Item",
        "tags": [
          "items"
        ]
      },
      "patch": {
        "description": "Update an item's fields, and its shares if the split changed.",
        "operationId": "update_item_api_v1_trips__slug__items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update Item",
        "tags": [
          "items"
        ]
      }
    },
...
  }
}